Separate modules for server and middleware.
Create separate modules for the handlers that interact with the
database. events_handlers interact with the events table to enable the
core realtime collaboration function. users_handlers enable markers on
the workspace.
