+++ blockly-rtc/server/Database.js
@@ -234,4 +234,4 @@ class Database {
   };
 };
 
-module.exports.Database = Database;
+module.exports = new Database();
